+++ climpred/tests/test_stats.py
@@ -16,7 +16,7 @@ from climpred.stats import (
     varweighted_mean_period,
 )
 
-BOOTSTRAP = 5
+ITERATIONS = 5
 
 
 def test_rm_trend_missing_dim():
@@ -120,7 +120,7 @@ def test_corr(PM_da_control_3d):
 
 def test_bootstrap_dpp_sig50_similar_dpp(PM_da_control_3d):
     sig = 50
-    actual = dpp_threshold(PM_da_control_3d, bootstrap=BOOTSTRAP, sig=sig).drop_vars(
+    actual = dpp_threshold(PM_da_control_3d, iterations=ITERATIONS, sig=sig).drop_vars(
         'quantile'
     )
     expected = dpp(PM_da_control_3d)
@@ -130,7 +130,7 @@ def test_bootstrap_dpp_sig50_similar_dpp(PM_da_control_3d):
 def test_bootstrap_vwmp_sig50_similar_vwmp(PM_da_control_3d):
     sig = 50
     actual = varweighted_mean_period_threshold(
-        PM_da_control_3d, bootstrap=BOOTSTRAP, sig=sig
+        PM_da_control_3d, iterations=ITERATIONS, sig=sig
     ).drop_vars('quantile')
     expected = varweighted_mean_period(PM_da_control_3d)
     xr.testing.assert_allclose(actual, expected, atol=2, rtol=0.5)
@@ -138,7 +138,7 @@ def test_bootstrap_vwmp_sig50_similar_vwmp(PM_da_control_3d):
 
 def test_bootstrap_func_multiple_sig_levels(PM_da_control_3d):
     sig = [5, 95]
-    actual = dpp_threshold(PM_da_control_3d, bootstrap=BOOTSTRAP, sig=sig)
+    actual = dpp_threshold(PM_da_control_3d, iterations=ITERATIONS, sig=sig)
     assert actual['quantile'].size == len(sig)
     assert (actual.isel(quantile=0).values <= actual.isel(quantile=1)).all()